Implement password vault CRUD operations

+++ db/sqlc/user.sql.go
@@ -11,7 +11,7 @@ import (
 )
 
 const createUser = `-- name: CreateUser :one
-INSERT INTO users(
+INSERT INTO users (
     username,
     email,
     password_hash,
@@ -80,7 +80,7 @@ SELECT id, username, email, password_hash, url, notes, icon, created_at, updated
 WHERE id = $1 LIMIT 1
 `
 
-func (q *Queries) GetUserByID(ctx context.Context, id int32) (User, error) {
+func (q *Queries) GetUserByID(ctx context.Context, id int64) (User, error) {
 	row := q.db.QueryRowContext(ctx, getUserByID, id)
 	var i User
 	err := row.Scan(
